libs/visualbase/pyproject.toml
[project]
name = "visualbase"
version = "0.1.0"
description = "Media platform for frame streaming"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]
zmq = [
    "pyzmq>=25.0.0",
]
webrtc = [
    "aiortc>=1.6.0",
    "aiohttp>=3.9.0",
]

[project.scripts]
visualbase = "visualbase.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/visualbase"]
